tasklist

+++ app/Http/Controllers/TasksController.php
@@ -17,14 +17,12 @@ class TasksController extends Controller
     // getでmessages/にアクセスされた場合の「一覧表示処理」
     public function index()
     {   
-        
-            $tasks = Task::orderBy('id', 'desc')->paginate(25);
+             $tasks = Task::orderBy('id', 'desc')->paginate(25);
             
-         return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+             return view('tasks.index', [
+                 'tasks' => $tasks,
+            ]);
             
-        
     }
 
     /**
@@ -57,10 +55,11 @@ class TasksController extends Controller
         ]);
         
         $task = new Task;
+        $task->user_id = $request->user()->id;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
-
+       
         return redirect('/');
     }
 
@@ -73,11 +72,18 @@ class TasksController extends Controller
     // getでmessages/idにアクセスされた場合の「取得表示処理」
     public function show($id)
     {
-        $task = Task::find($id);
+        $task = \App\Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+                $task = Task::find($id);
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+                return view('tasks.show', [
+                'task' => $task,
+            ]);
+            return redirect('/');
+        }
+        
+        return redirect('/');
     }
 
     /**
@@ -88,12 +94,20 @@ class TasksController extends Controller
      */
     // getでmessages/id/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
-    {
-        $task = Task::find($id);
+    {   
+        
+        $task = \App\Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+            $task = Task::find($id);
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        
+           return redirect('/');
+        }
+        return redirect('/');
     }
 
     /**
@@ -128,10 +142,15 @@ class TasksController extends Controller
     // deleteでmessages/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
+        $task = \App\Task::find($id);
         
-        $task = Task::find($id);
-        $task->delete();
-
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task = Task::find($id);
+            $task->delete();
+            
+            return redirect('/');
+        }
+         
+         return redirect('/');
     }
 }
